backend/requirements.txt
asgiref==3.7.2
certifi==2023.11.17
charset-normalizer==3.3.2
Django==4.1.7
django-arrayfield==1.0.0rc0
django-cors-headers==4.3.1
django-debug-toolbar==4.0.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
idna==3.6
mysqlclient==2.2.1
PyJWT==2.8.0
python-dotenv==1.0.0
pytz==2023.3.post1
requests==2.31.0
sqlparse==0.4.4
urllib3==2.1.0
